+++ src/mqtt/qmqtttopicfilter.h
@@ -43,7 +43,7 @@ class QMqttTopicFilterPrivate;
 
 class QMqttTopicFilter;
 // qHash is a friend, but we can't use default arguments for friends (§8.3.6.4)
-Q_MQTT_EXPORT uint qHash(const QMqttTopicFilter &name, uint seed = 0) Q_DECL_NOTHROW;
+Q_MQTT_EXPORT size_t qHash(const QMqttTopicFilter &name, size_t seed = 0) Q_DECL_NOTHROW;
 
 class Q_MQTT_EXPORT QMqttTopicFilter
 {
@@ -77,7 +77,7 @@ public:
     friend Q_MQTT_EXPORT bool operator==(const QMqttTopicFilter &lhs, const QMqttTopicFilter &rhs) Q_DECL_NOTHROW;
     friend inline bool operator!=(const QMqttTopicFilter &lhs, const QMqttTopicFilter &rhs) Q_DECL_NOTHROW { return !(lhs == rhs); }
     friend Q_MQTT_EXPORT bool operator<(const QMqttTopicFilter &lhs, const QMqttTopicFilter &rhs) Q_DECL_NOTHROW;
-    friend Q_MQTT_EXPORT uint qHash(const QMqttTopicFilter &filter, uint seed) Q_DECL_NOTHROW;
+    friend Q_MQTT_EXPORT size_t qHash(const QMqttTopicFilter &filter, size_t seed) Q_DECL_NOTHROW;
 
 private:
     QExplicitlySharedDataPointer<QMqttTopicFilterPrivate> d;
